requirements.txt
pyYAML
ipython
